split out 'output'
