+++ core/incremental/ddTests.ts
@@ -1,6 +1,6 @@
 import { toGraphviz } from "./graphviz";
-import { Rec, Statement } from "../types";
-import { formatOutput, IncrementalInterpreter } from "./interpreter";
+import { Rec } from "../types";
+import { IncrementalInterpreter } from "./interpreter";
 import { fsLoader } from "../fsLoader";
 import { Suite } from "../../util/testBench/testing";
 import { ProcessFn, runDDTestAtPath, TestOutput } from "../../util/ddTest";
@@ -15,6 +15,7 @@ import {
   parserTermToInternal,
 } from "../translateAST";
 import { getJoinInfo } from "./build";
+import { formatOutput } from "./output";
 
 export function incrTests(writeResults: boolean): Suite {
   const tests: [string, ProcessFn][] = [
