+++ core/incremental/interpreter.ts
@@ -1,4 +1,9 @@
-import { EmissionLogAndGraph, emptyRuleGraph, RuleGraph } from "./types";
+import {
+  EmissionLogAndGraph,
+  emptyRuleGraph,
+  Output,
+  RuleGraph,
+} from "./types";
 import { baseFactTrace, Res, Rule, Statement } from "../types";
 import { doQuery, EmissionLog, insertFact } from "./eval";
 import { ppb, ppr } from "../pretty";
@@ -18,12 +23,6 @@ import {
   emptyCatalog,
 } from "./catalog";
 import { buildGraph } from "./build";
-
-export type Output =
-  | { type: "EmissionLog"; log: EmissionLog }
-  | { type: "Trace"; logAndGraph: EmissionLogAndGraph }
-  | { type: "QueryResults"; results: Res[] }
-  | { type: "Acknowledge" };
 
 export class IncrementalInterpreter extends AbstractInterpreter {
   graph: RuleGraph | null;
@@ -153,54 +152,3 @@ function replayFacts(ruleGraph: RuleGraph, catalog: Catalog): RuleGraph {
 }
 
 const ack: Output = { type: "Acknowledge" };
-
-type OutputOptions = {
-  emissionLogMode: "test" | "repl";
-  showBindings: boolean;
-};
-
-export function formatOutput(
-  graph: RuleGraph,
-  output: Output,
-  opts: OutputOptions
-): TestOutput {
-  switch (output.type) {
-    case "Acknowledge":
-      return datalogOut([]);
-    case "EmissionLog":
-      if (opts.emissionLogMode === "test") {
-        return {
-          mimeType: "incremental-datalog/trace",
-          content: output.log
-            .map(
-              ({ fromID, output }) =>
-                `${fromID}: [${output
-                  .map((res) => (res.term ? ppr(res) : ppb(res.bindings)))
-                  .join(", ")}]`
-            )
-            .join("\n"),
-        };
-      } else {
-        return datalogOut(
-          flatMap(
-            output.log.filter((emissionBatch) => {
-              const fromNode = graph.nodes.get(emissionBatch.fromID);
-              return (
-                !fromNode.isInternal && fromNode.desc.type !== "BaseFactTable"
-              );
-            }),
-            ({ output }) => output.map((res) => res.term)
-          )
-        );
-      }
-    case "QueryResults":
-      return opts.showBindings
-        ? datalogOutResults(output.results)
-        : datalogOut(output.results.map((res) => res.term));
-    case "Trace":
-      return {
-        content: JSON.stringify(output.logAndGraph),
-        mimeType: "incremental-datalog/trace",
-      };
-  }
-}
